app/requirements.txt
Flask<3
mysql-connector-python
